examples/twilio-chatbot/requirements.txt
git+https://github.com/chillmediator/pipecat.git@main#egg=pipecat[cartesia,openai,silero,deepgram]
fastapi
uvicorn
python-dotenv
loguru
fsspec[gdrive]>=2023.12.0
aiofiles>=23.2.1
